Move extension UTs to separate suite
